refactor(docs): move the snapshot budget into a testable module
process.exit can truncate a piped stderr write, so the budget failure sets exitCode and returns instead.
